requirements.txt
# Core chatbot UI
chainlit>=0.7.0

# Manifest LLM orchestrator
manifest-ml[server]>=0.0.19

# OpenAI support (used by Manifest or directly)
openai>=1.0.0

# HTTP client for local/ollama
httpx>=0.27.0

# Optional: Hugging Face Transformers (for local inference)
transformers>=4.35.0
torch>=2.0.0

# Optional: Requests for direct Hugging Face API calls (if needed)
requests>=2.28.0

# Optional: SQLite/Redis support if you enable cache in Manifest
aiosqlite>=0.18.0         # for sqlite cache
redis>=4.6.0              # for redis cache
